Move reapi structures into trait implementation
